==> Script 3 - IATI (non-linked) partner activities_kp.R <==
# --------------------------------------------------------------- #
# Script 3
# Extract ODA research activities from public IATI data #
# --------------------------------------------------------------- #

### 1) Extract IATI data on RI partner activities (linked or manually identified) ----

# Read in linked partner IATI activity info from script 2
ri_linked_activites <- readRDS(file = "Outputs/ri_linked_activites_Jan24update.rds")

# Manually add on other (non-linked) partner activities from Excel
partner_iati_activity_ids <- unlinked_partner_iati_activity_ids %>% 
  plyr::rbind.fill(ri_linked_activites)


specific_activity_extract <- function(activity_id) {
  page_list <- list()
  page <- 1
  page_df = data.frame()
  
  while (!is.null(page_df)) {
    Sys.sleep(1)
    message(page)
    page_df <- iati_activity_extract(page, activity_id)
    if(!is.null(page_df)){
      page_list[[page]] = page_df
    }
    page = page + 1
  }
  rbindlist(page_list, fill=T)
}

uk_activity_ids = unique(partner_iati_activity_ids$iati_identifier)
uk_activity_ids <- gsub(" ", "%", uk_activity_ids)
uk_activity_ids_new <- URLencode(ri_linked_activites_new$iati_identifier)

uk_activity_ids <- uk_activity_ids[!grepl("/|)", uk_activity_ids)]
uk_activity_ids_symbols <- uk_activity_ids[grepl("/|)", uk_activity_ids)]
# uk_activity_ids_p1 <- uk_activity_ids[218]

uk_activity_ids <- str_replace(uk_activity_ids, "CA-CRA_ACR-811793611-%CA-CRA_ACR-811793611-HGC_JUN2022", "CA-CRA_ACR-811793611-HGC")
uk_activity_ids <- str_replace(uk_activity_ids, "GB-COH-03122495-A0268DFID%Programme%Directorate%education", "GB-COH-03122495-A0268DFID%PROGRAMME%DIRECTORATE%EDUCATION")
uk_activity_ids <- uk_activity_ids[!grepl("GB-COH-03122495-A0268DFID%PROGRAMME%DIRECTORATE%EDUCATION", uk_activity_ids)]

# i think below code will work, just it doesn't like certain symbols such as brackets
# need to find a way to remove or get those to run

partner_activity_extract <- lapply(uk_activity_ids, specific_activity_extract)
partner_activity_extract = rbindlist(partner_activity_extract, fill=T)
# partner_activity_extract_new <- lapply(uk_activity_ids_new, specific_activity_extract)
# partner_activity_extract_new = rbindlist(partner_activity_extract_new, fill=T)

# batch_size = 15
# batches = c()
# current_batch = c()
# for(i in 1:length(uk_activity_ids)){
#   current_id = uk_activity_ids[i]
#   if(i %% batch_size == 0){
#     current_batch_str = paste0('("', paste(current_batch, collapse = '" OR "'), '")')
#     batches = c(batches, current_batch_str)
#     current_batch = c(current_id)
#   } else {
#     current_batch = c(current_batch, current_id)
#   }
# }
# 
# partner_activity_extract <- lapply(batches, specific_activity_extract)
# partner_activity_extract = rbindlist(partner_activity_extract, fill=TRUE)

#partner_activity_extract <- lapply(unique(partner_iati_activity_ids$iati_identifier), specific_activity_extract)
#partner_activity_extract = rbindlist(partner_activity_extract, fill=T)


# length(unique(partner_activity_extract$reporting_org_ref))
# unique(partner_activity_extract$activity_id)
# unique(partner_activity_extract$gov_funder)
# nrow(partner_activity_extract[partner_activity_extract$gov_funder=="Foreign, Commonwealth and Development Office",])
# partner_activity_extract[partner_activity_extract$reporting_org_ref=="GB-COH-877338",]

#       # Prepare results data frame and counters
#       partner_activity_extract <- data.frame()
# 
#       # Run extraction, stopping when no new sector codes returned
#       for (id in partner_iati_activity_ids$iati_id) {
# 
#           print(id)
#           result <- iati_activity_extract(id)
#           partner_activity_extract <- c(partner_activity_extract, result)
# 
#       }
# 
# partner_activity_extract <- bind_rows(partner_activity_extract)

# partner_activity_extract <- rbind(partner_activity_extract,partner_activity_extract_new, fill=TRUE)

# Save to Rdata file
# saveRDS(partner_activity_extract, file = "Outputs/partner_activity_extract_Jan24update.rds")
partner_activity_extract <- readRDS(file = "Outputs/partner_activity_extract_Jan24update.rds")
# partner_activity_extract_kp <- readRDS(file = "Outputs/partner_activity_extract_kp.rds")

# Join on funder and fund information
partner_activity_extract <- partner_activity_extract %>%       
      left_join(partner_iati_activity_ids, by = "iati_identifier") %>% 
      select(-extending_org)
# %>% filter(!is.na(activity_id))

### 2) Activity extract for specific partnership organisations ----

# FCDO (part-)funded partnership activities 

# having problems with multiple languages for IDRC, wont unnest properly when looking at participating organisation
# org_code <- c(
#               "XM-DAC-47015", # CGIAR
#               "XM-DAC-301-2", # IDRC
#               "DAC-1601",     # Bill & Melinda Gates Foundation
#               "XI-IATI-AGR"   # AgResults (Consortium)
#               )   

# temporarily removing IDRC from the org_code list
org_code <- c(
  "XM-DAC-47015", # CGIAR
  "DAC-1601",     # Bill & Melinda Gates Foundation
  "XI-IATI-AGR"   # AgResults (Consortium)
)


# 1) Activity extract

specific_org_extract <- function(o_code) {
  page_list <- list()
  page <- 1
  page_df = data.frame()
  
  while (!is.null(page_df)) {
    Sys.sleep(1)
    message(page)
    page_df <- org_activity_extract(page, o_code)
    if(!is.null(page_df)){
      page_list[[page]] = page_df
    }
    page = page + 1
  }
  rbindlist(page_list, fill=T)
}

# specific_org_extract("GB-GOV-15")

org_activity_list <- lapply(org_code, specific_org_extract)
org_activity_list = rbindlist(org_activity_list, fill=T)


# 1a) Extracting IDRC separately with slightly different function to only extract the "en" participating org names

IDRC <- "XM-DAC-301-2"

specific_org_extract_IDRC <- function(o_code) {
  page_list <- list()
  page <- 1
  page_df = data.frame()
  
  while (!is.null(page_df)) {
    Sys.sleep(1)
    message(page)
    page_df <- org_activity_extract_IDRC(page, o_code)
    if(!is.null(page_df)){
      page_list[[page]] = page_df
    }
    page = page + 1
  }
  rbindlist(page_list, fill=T)
}

org_activity_list_IDRC <- lapply(IDRC, specific_org_extract_IDRC)
org_activity_list_IDRC = rbindlist(org_activity_list_IDRC, fill=T)
org_activity_list_IDRC <- org_activity_list_IDRC %>% filter(participating_org_narrative_xml_lang == "en")

org_activity_list_IDRC$participating_org_ref <- ifelse(org_activity_list_IDRC$participating_org_narrative %in% c("International Development Research Centre","Canada. Parliament"), "XM-DAC-301-2", "")
org_activity_list_IDRC$participating_org_ref <- ifelse(org_activity_list_IDRC$participating_org_narrative %in% c("Bill & Melinda Gates Foundation"), "DAC-1601", org_activity_list_IDRC$participating_org_ref)
org_activity_list_IDRC$participating_org_ref <- ifelse(org_activity_list_IDRC$participating_org_narrative %in% c("Norwegian Agency for Development Cooperation"), "NO-BRC-971277882", org_activity_list_IDRC$participating_org_ref)


org_activity_list <- plyr::rbind.fill(org_activity_list,org_activity_list_IDRC)


# saveRDS(org_activity_list, "Outputs/org_activity_list_Jan24update.rds")
org_activity_list <- readRDS("Outputs/org_activity_list_Jan24update.rds")


        # 2.A) Unlist activity titles and subset for those that mention FCDO/DFID
        # (identifies FCDO-funded Gates Foundation activities)

        partner_activities_via_title <- org_activity_list %>%
          filter(reporting_org_ref == "DAC-1601") %>%  # Gates org ID
          filter(str_detect(title_narrative, "FCDO|DFID")) %>%
          mutate(gov_funder = "Foreign, Commonwealth and Development Office",
                 fund = "FCDO Research - Partnerships") %>%
          select(iati_identifier, gov_funder, fund) %>%
          unique()

        # 2.B) Identify UK gov funded activities from participating organisations

        # below does not use "GB-GOV-1" to search for programmes but includes more words to search the narratives
        
        partner_activities_via_funder <- org_activity_list %>%
          # restrict to UK gov funding
          select(iati_identifier, participating_org_role, participating_org_narrative, participating_org_ref, participating_org_activity_id) %>%
          unnest(cols = c(participating_org_role, participating_org_narrative, participating_org_ref,participating_org_activity_id)) %>%
          filter(str_detect(participating_org_role, "1")|str_detect(iati_identifier, "XI-IATI-AGR")) %>%
          filter(str_detect(participating_org_narrative, "Britain|DFID|Department for International Development|Foreign, Commonwealth|FCDO|DHSC|Department of Health and Social Care") |
                   str_detect(iati_identifier, "DFID|FCDO") |
                   str_detect(iati_identifier, "XI-IATI-AGR")      # AgResults partially funded
          ) %>%
          mutate(gov_funder = if_else(str_detect(participating_org_narrative, "Health"), "Department of Health and Social Care",
                                      "Foreign, Commonwealth and Development Office"),
                 fund = case_when(
                   # IDRC GAMRIF projects
                   str_detect(iati_identifier, "XM-DAC-301-2") & str_detect(participating_org_narrative, "Health") ~ "DHSC - Global Health Security - GAMRIF",
                   # Other DHSC partnerships
                   str_detect(participating_org_narrative, "Health") ~ "DHSC - Global Health Research - Partnerships",
                   # FCDO funding
                   TRUE ~ "FCDO Research - Partnerships"
                 )) %>%
          select(iati_identifier, gov_funder, fund) %>%
          unique()
        

#         partner_activities_via_funder2_unnest <- partner_activities_via_funder2 %>% unnest(cols = c(participating_org_ref))
#         unique(partner_activities_via_funder2_unnest$participating_org_ref)
        
        # participating_org_id <- org_activity_list %>%
        #   select(iati_identifier, participating_org_role, participating_org_activity_id) %>%
        #   unnest(cols = c(participating_org_role, participating_org_activity_id)) %>%
        #   filter(participating_org_role == "1" & !is.na(participating_org_activity_id)) %>%
        #   unique()
        # 
        # participating_org_id <- participating_org_id %>%
        #   select(iati_identifier,participating_org_activity_id) %>%
        #   filter(participating_org_activity_id != "")
         
        # Combine 2A and 2B
        partnership_activities <- plyr::rbind.fill(partner_activities_via_title, partner_activities_via_funder)
        
        # Filter original data
        partnership_activities <- org_activity_list %>%
          inner_join(partnership_activities, by = "iati_identifier")

# Save to Rdata file
# saveRDS(partnership_activities, file = "Outputs/partnership_activities_Jan24update.rds")
partnership_activities <- readRDS(file = "Outputs/partnership_activities_Jan24update.rds")
# partnership_activities_ec <- readRDS(file = "Outputs/partnership_activities.rds")

### C) Combine individual with partner activities (extractions A and B above) ---- 

partner_activity_extract <- partner_activity_extract %>% filter(!iati_identifier %in% c(partnership_activities$iati_identifier))

partner_activity_comb <- plyr::rbind.fill(as.data.frame(partner_activity_extract), as.data.frame(partnership_activities))
# partner_activity_comb <- plyr::rbind.fill(as.data.frame(partner_activity_extract), as.data.frame(partnership_activities)) %>%
#   filter(default_flow_type_code != "50")

# partner_activity_comb$activity_id <- ifelse(partner_activity_comb$activity_id %in% c("GB-COH-RC000658-GB-GOV-1-301132","cGB-COH-RC000658-GB-GOV-1-301132"),
#                                            "GB-COH-RC000658 -GB-GOV-1-301132", partner_activity_comb$activity_id)

partner_activity_comb <- partner_activity_comb %>% filter(iati_identifier!="US-EIN-134166228-GB-COH-06274284-TEA2 - Pioneer Energy Investment Initiative 2")

# D) Extract nested activity data ----------------------------------------------

# Extract base activity information - hierarchy and status
activity_list_base <- partner_activity_comb %>% 
  select(iati_identifier, hierarchy, 
         activity_status = activity_status_code,
         gov_funder,
         activity_id,
         fund) %>% 
  unique() %>%
  mutate(activity_status = str_replace(activity_status, "1", "Pipeline/identification"),
         activity_status = str_replace(activity_status, "2", "Implementation"),
         activity_status = str_replace(activity_status, "3", "Finalisation"),
         activity_status = str_replace(activity_status, "4", "Closed"),
         activity_status = str_replace(activity_status, "5", "Cancelled"),
         activity_status = str_replace(activity_status, "6", "Suspended"))


# 1) Unlist activity title and description
activity_list_unnest_1_title <- partner_activity_comb %>%
  select(iati_identifier, title_narrative, title_narrative_xml_lang) %>%
  unnest(cols = c(title_narrative,title_narrative_xml_lang)) %>%
  mutate(title_narrative_xml_lang = ifelse(is.na(title_narrative_xml_lang),"",title_narrative_xml_lang)) %>%
filter(title_narrative_xml_lang != "fr") %>%
  select(iati_identifier, title_narrative) %>%
  unique() %>% 
  group_by(iati_identifier) %>%
  summarise(title_narrative = paste(coalesce(title_narrative, ""), collapse = "; ")) %>%
  rename(activity_title = title_narrative)

# need to get below to run - more problems with french names!!

# need to identify where names and languages don't match

activity_list_unnest_1_description <- partner_activity_comb %>%
  select(iati_identifier, description_narrative, description_narrative_xml_lang)
activity_list_unnest_1_description$len_narr <- lengths(activity_list_unnest_1_description$description_narrative)
activity_list_unnest_1_description$len_lang <- lengths(activity_list_unnest_1_description$description_narrative_xml_lang)
activity_list_unnest_1_description$diff <- activity_list_unnest_1_description$len_narr - activity_list_unnest_1_description$len_lang

activity_list_unnest_1_description_with_lists <- activity_list_unnest_1_description[activity_list_unnest_1_description$len_narr > 1,]
activity_list_unnest_1_description_with_lists <- activity_list_unnest_1_description_with_lists[activity_list_unnest_1_description_with_lists$len_lang != 0,]
activity_list_unnest_1_description_with_lists <- activity_list_unnest_1_description_with_lists[activity_list_unnest_1_description_with_lists$diff != 0,]
activity_list_unnest_1_description_with_lists <- activity_list_unnest_1_description_with_lists$iati_identifier

# now unnesting everything that isn't those that mismatch

activity_list_unnest_1_description1 <- activity_list_unnest_1_description[!activity_list_unnest_1_description$iati_identifier %in% c(activity_list_unnest_1_description_with_lists),]

activity_list_unnest_1_description1 <- activity_list_unnest_1_description1 %>%
  select(iati_identifier, description_narrative, description_narrative_xml_lang) %>%
  unnest(cols = c(description_narrative,description_narrative_xml_lang)) %>%
  mutate(description_narrative_xml_lang = ifelse(is.na(description_narrative_xml_lang),"",description_narrative_xml_lang)) %>%
  filter(description_narrative_xml_lang != "fr") %>%
  select(iati_identifier, description_narrative) %>%
  unique() %>% 
  group_by(iati_identifier) %>%
  summarise(description_narrative = paste(coalesce(description_narrative, ""), collapse = "; ")) %>%
  rename(activity_description = description_narrative)

activity_list_unnest_1_description2 <- activity_list_unnest_1_description[activity_list_unnest_1_description$iati_identifier %in% c(activity_list_unnest_1_description_with_lists),]

activity_list_unnest_1_description2 <- activity_list_unnest_1_description2 %>%
  select(iati_identifier, description_narrative) %>%
  unnest(cols = c(description_narrative)) %>%
  filter(row_number() %% 2 == 1) %>%
  unique() %>% 
  group_by(iati_identifier) %>%
  summarise(description_narrative = paste(coalesce(description_narrative, ""), collapse = "; ")) %>%
  rename(activity_description = description_narrative)

activity_list_unnest_1 <- full_join(activity_list_unnest_1_title, activity_list_unnest_1_description1, activity_list_unnest_1_description2, by = "iati_identifier")

# Summarise records with multiple "General" descriptions
# activity_list_unnest_1 <- activity_list_unnest_1 %>%
#       group_by(iati_identifier, title_narrative, reporting_org_type) %>%
#       summarise(text = paste(coalesce(text, ""), collapse = "\n\n")) %>%
#       spread(key = reporting_org_type, value = text) %>%
#       mutate(activity_description = if_else(!is.na(Objectives), paste0(General, "\n\n", Objectives), General)) %>%
#       ungroup()


# 2) Unlist recipient countries

# activity_list_unnest_2 <- partner_activity_comb %>% 
#   filter(lengths(recipient_country) != 0) %>% 
#   unnest(cols = recipient_country) %>% 
#   select(iati_identifier, country.name) %>% 
#   group_by(iati_identifier) %>%
#   unique() %>% 
#   summarise(country_name = paste(coalesce(country.name, ""), collapse = ", "))

activity_list_unnest_2_country <- partner_activity_comb %>%
  #filter(lengths(recipient_country_code) != 0) %>%
  unnest(c(recipient_country_code)) %>% 
  select(iati_identifier, recipient_country_code)

activity_list_unnest_2_country$recipient_country <- countrycode_list$name[match(activity_list_unnest_2_country$recipient_country_code,countrycode_list$code)]

activity_list_unnest_2_country <- activity_list_unnest_2_country %>% select(-recipient_country_code) %>%
  unique() %>% 
  group_by(iati_identifier) %>%
  summarise(recipient_country = paste(coalesce(recipient_country, ""), collapse = ", "))


activity_list_unnest_2_region <- partner_activity_comb %>%
  #filter(lengths(recipient_country_code) != 0) %>%
  unnest(c(recipient_region_code)) %>% 
  select(iati_identifier, recipient_region_code)

activity_list_unnest_2_region$recipient_region <- regioncode_list$name[match(activity_list_unnest_2_region$recipient_region_code,regioncode_list$code)]

activity_list_unnest_2_region <- activity_list_unnest_2_region %>% select(-recipient_region_code) %>%
  unique() %>% 
  group_by(iati_identifier) %>%
  summarise(recipient_region = paste(coalesce(recipient_region, ""), collapse = ", "))

activity_list_unnest_2 <- full_join(activity_list_unnest_2_country, activity_list_unnest_2_region, by = "iati_identifier")

activity_list_unnest_2$recipient_country_region <- coalesce(activity_list_unnest_2$recipient_country, activity_list_unnest_2$recipient_region)

activity_list_unnest_2 <- activity_list_unnest_2 %>% 
  mutate(across(c(recipient_country_region), na_if, "Developing countries, unspecified")) %>%
  select(-recipient_country, -recipient_region)

  
# 3) Unlist sectors
# activity_list_unnest_3 <- partner_activity_comb %>% 
#   filter(lengths(sector) != 0) %>% 
#   unnest(cols = sector) %>% 
#   select(iati_identifier, sector.name) %>% 
#   filter(sector.name != "Vocabulary 99 or 98") %>% 
#   group_by(iati_identifier) %>%
#   unique() %>% 
#   summarise(sector_name = paste(coalesce(sector.name, ""), collapse = ", ")) %>% 
#   ungroup()

activity_list_unnest_3 <- partner_activity_comb %>%
  #filter(lengths(sector_code) != 0) %>%
  unnest(cols = sector_code) %>% 
  select(iati_identifier, sector_code) %>% 
  unique()

activity_list_unnest_3$sector_name <- sector_list$name[match(activity_list_unnest_3$sector_code,sector_list$code)]

activity_list_unnest_3 <- activity_list_unnest_3 %>% select(-sector_code) %>%
  unique() %>%
  group_by(iati_identifier) %>%
  summarise(sector_name = paste(coalesce(sector_name, ""), collapse = ", "))


# activity_list_unnest_3[activity_list_unnest_3$iati_identifier=="GB-COH-03122495-A0014",]

# 4) Unlist implementing organisations

# code below is trying to work out where the mismatching names/roles are

activity_list_unnest_4 <- partner_activity_comb %>%
  select(iati_identifier, participating_org_role, participating_org_narrative, participating_org_ref)
activity_list_unnest_4$len_role <- lengths(activity_list_unnest_4$participating_org_role)
activity_list_unnest_4$len_narr <- lengths(activity_list_unnest_4$participating_org_narrative)
activity_list_unnest_4$len_ref <- lengths(activity_list_unnest_4$participating_org_ref)
activity_list_unnest_4$rolenarr_diff <- activity_list_unnest_4$len_role - activity_list_unnest_4$len_narr
activity_list_unnest_4$narrref_diff <- activity_list_unnest_4$len_narr - activity_list_unnest_4$len_ref
activity_list_unnest_4$roleref_diff <- activity_list_unnest_4$len_role - activity_list_unnest_4$len_ref

activity_list_unnest_4_orgdiffs <- activity_list_unnest_4[activity_list_unnest_4$rolenarr_diff != 0,]
activity_list_unnest_4 <- activity_list_unnest_4[!activity_list_unnest_4$rolenarr_diff != 0,]

activity_list_unnest_4_1 <- activity_list_unnest_4 %>%
  select(iati_identifier, participating_org_role, participating_org_narrative) %>%
  unnest(c(participating_org_role, participating_org_narrative)) %>%
  filter(participating_org_role == "4") %>%
  unique()

activity_list_unnest_4_1$participating_org_narrative <- ifelse(activity_list_unnest_4_1$participating_org_narrative == "**********",
                                                               "", activity_list_unnest_4_1$participating_org_narrative)


activity_list_unnest_4_orgdiffs <- activity_list_unnest_4_orgdiffs %>%
  mutate(participating_org_role = map2(participating_org_role, len_narr, ~head(.x, .y)))

activity_list_unnest_4_2 <- activity_list_unnest_4_orgdiffs %>%
  select(iati_identifier, participating_org_role, participating_org_narrative) %>%
  unnest(c(participating_org_role, participating_org_narrative)) %>%
  filter(participating_org_role == "4") %>%
  unique()

activity_list_unnest_4 <- rbind(activity_list_unnest_4_1, activity_list_unnest_4_2)
rm(activity_list_unnest_4_1,activity_list_unnest_4_2,activity_list_unnest_4_orgdiffs)

activity_list_unnest_4_refs <- partner_activity_comb %>%
  select(iati_identifier, participating_org_role, participating_org_narrative, participating_org_ref)

# activity_list_unnest_4_refs <- activity_list_unnest_4_refs %>%
#   mutate(participating_org_narrative = map2(participating_org_narrative, len_ref, ~head(.x, .y))) %>%
#   select(participating_org_narrative, participating_org_ref) %>%
#   unnest(c(participating_org_narrative, participating_org_ref)) %>%
#   unique()

activity_list_unnest_4_refs$len_narr <- lengths(activity_list_unnest_4_refs$participating_org_narrative)
activity_list_unnest_4_refs$len_ref <- lengths(activity_list_unnest_4_refs$participating_org_ref)
activity_list_unnest_4_refs$narrref_diff <- activity_list_unnest_4_refs$len_narr - activity_list_unnest_4_refs$len_ref
mismatch <- activity_list_unnest_4_refs[activity_list_unnest_4_refs$narrref_diff != 0,]
mismatch_narr <- unlist(mismatch$participating_org_narrative)
mismatch_ref <- unlist(mismatch$participating_org_ref)
mismatch_ref <- mismatch_ref[-1]
mismatch <- data.frame(mismatch_narr, mismatch_ref)

activity_list_unnest_4_refs <- activity_list_unnest_4_refs %>%
  filter(narrref_diff == 0) %>%
  select(participating_org_narrative, participating_org_ref) %>%
  unnest(c(participating_org_narrative, participating_org_ref)) %>%
  unique()


names(mismatch) <- names(activity_list_unnest_4_refs)

activity_list_unnest_4_refs <- rbind(activity_list_unnest_4_refs, mismatch)


#"**********"
#"Not registered in IATI"

# Will need to run the code separately for IDRC using Alex from IATI new bit of code

# activity_list_unnest_4_narr <- partner_activity_comb %>%
#   select(iati_identifier, participating_org_narrative_xml_lang, participating_org_narrative) %>%
#   unnest(cols = c(participating_org_narrative_xml_lang,participating_org_narrative)) %>%
#   mutate(participating_org_narrative_xml_lang = ifelse(is.na(participating_org_narrative_xml_lang),"",participating_org_narrative_xml_lang)) %>%
#   filter(!participating_org_narrative_xml_lang %in% c("fr","es")) %>%
#   select(-participating_org_narrative_xml_lang) %>%
#   group_by(iati_identifier) %>%
#   summarise(participating_org_narrative = paste(coalesce(participating_org_narrative, ""), collapse = "; "))
# 
# activity_list_unnest_4_role <- partner_activity_comb %>% 
#   #filter(lengths(participating_org_ref) != 0) %>%
#   #unnest(c(participating_org_ref, participating_org_narrative)) %>% 
#   select(iati_identifier, participating_org_role) %>% 
#   unnest(c(participating_org_role)) %>%
#   group_by(iati_identifier) %>%
#   summarise(participating_org_role = paste(coalesce(participating_org_role, ""), collapse = "; "))
# 
# activity_list_unnest_4 <- activity_list_unnest_4_role %>% left_join(activity_list_unnest_4_narr, by = "iati_identifier")


  # Identify activities with no implementing partner info
    no_partner_info <- activity_list_unnest_4 %>%
      select(iati_identifier) %>%
      mutate(has_implementing_partner_info = "Yes") %>%
      unique() %>%
      right_join(partner_activity_comb, by = "iati_identifier") %>%
      filter(is.na(has_implementing_partner_info)) %>%
      select(iati_identifier) %>%
      unique()
  
  # Add country locations based on IATI org references or lookup
  # (takes ~5 mins to run)
    # below code not working due to problems with org ref id not matching narrative
    
# activity_list_unnest_4 <- activity_list_unnest_4 %>%
#     # Extract 2 digit country code from org references (where populated)
#     mutate(country_code = if_else((!is.na(participating_org_ref) & substr(participating_org_ref,3,3) == "-" & !(substr(participating_org_ref,1,2) %in% c("XI", "XM"))), 
#                                   substr(participating_org_ref,1,2), ""))
  
  # Function searches the org title (e.g. Kenya) and returns a country if its present in the name
activity_list_unnest_4$country_name2 <- map(activity_list_unnest_4$participating_org_narrative, org_country_lookup)
activity_list_unnest_4$country_name2 <- as.character(activity_list_unnest_4$country_name2)
  activity_list_unnest_4$country_name2 <- ifelse(activity_list_unnest_4$country_name2 == "NA", "", activity_list_unnest_4$country_name2)
  
  
  # check if input is a valid 2-digit country code and then does a lookup to the countrycode_list
  # activity_list_unnest_4$country_name <- ifelse(is.na(activity_list_unnest_4$country_code) | nchar(activity_list_unnest_4$country_code) < 2, NA, countrycode_list$name[match(activity_list_unnest_4$country_code,countrycode_list$code)])
  # activity_list_unnest_4$country_name <- ifelse(is.na(activity_list_unnest_4$country_name), "", activity_list_unnest_4$country_name)
  
  # activity_list_unnest_4[activity_list_unnest_4 == ""] <- NA
  
  # This needs fixing because currently I don't know why it's only picking one of the country names when there might be several 
  # activity_list_unnest_4$partner_country <- coalesce(activity_list_unnest_4$country_name, activity_list_unnest_4$country_name2)
  activity_list_unnest_4 <- activity_list_unnest_4 %>% rename(partner_country = country_name2)
  
  # partner_org_lookup <- activity_list_unnest_4 %>% select(participating_org_narrative, partner_country) %>% unique()
  # write.csv(partner_org_lookup, file = "Inputs/partner_org_lookup.csv", na = "")
  
  # activity_list_unnest_4 <- activity_list_unnest_4%>% select(-country_name, -country_name2)
  
  # Summarise partner org countries and names
  activity_list_unnest_4_countries <- activity_list_unnest_4 %>% 
    select(iati_identifier, partner_country) %>% 
    unique() %>% 
    filter(!is.na(partner_country)) %>% 
    group_by(iati_identifier) %>% 
    summarise(partner_country = paste(partner_country, collapse = ", "))
  
  activity_list_unnest_4_partners <- activity_list_unnest_4 %>% 
    select(iati_identifier, participating_org_narrative) %>% 
    unique() %>% 
    filter(!is.na(participating_org_narrative)) %>% 
    group_by(iati_identifier) %>% 
    summarise(partner = paste(coalesce(participating_org_narrative, ""), collapse = ", ")) 
  
  # Add partner name and country info to master dataset
  activity_list_unnest_4 <- activity_list_unnest_4 %>% 
    #select(-participating_org_narrative, -participating_org_ref, -country_code, -participating_org_role, -partner_country) %>%
    select(-participating_org_narrative, -participating_org_role, -partner_country) %>%
    left_join(activity_list_unnest_4_partners, by = "iati_identifier") %>% 
    left_join(activity_list_unnest_4_countries, by = "iati_identifier") %>%
    unique()

# 5) Unlist publishing org
  
  # activity_list_unnest_5 <- uk_gov_list_filtered %>% 
  #   #filter(lengths(participating_org_ref) != 0) %>%
  #   #unnest(c(participating_org_ref, participating_org_narrative)) %>% 
  #   select(iati_identifier, participating_org_role, participating_org_narrative) %>% 
  #   unnest(c(participating_org_role, participating_org_narrative)) %>% 
  #   filter(participating_org_role == "3") %>% 
  #   unique() %>%  select(-participating_org_role) %>% rename(extending_org = participating_org_narrative)
  
  activity_list_unnest_5 <- partner_activity_comb %>%
  filter(lengths(reporting_org_narrative) != 0) %>%
  unnest(cols = reporting_org_narrative) %>%
  select(iati_identifier,
         reporting_org_ref,reporting_org_narrative) %>%
  # take top (English) name in cases of different languages
  group_by(iati_identifier, reporting_org_ref) %>%
  slice(1) %>%
  unique() %>%
  ungroup()

    # Lookup country (takes ~10 mins to run)
    activity_list_unnest_5 <- activity_list_unnest_5 %>% 
        # Extract 2 digit country code from org references (where populated)
      mutate(country_code = if_else((!is.na(reporting_org_ref) & substr(reporting_org_ref,3,3) == "-" & !(substr(reporting_org_ref,1,2) %in% c("XI", "XM"))), 
                                    substr(reporting_org_ref,1,2), ""))
   
     # Function searches the org title (e.g. Kenya) and returns a country if its present in the name
    activity_list_unnest_5$country_name2 <- map(activity_list_unnest_5$reporting_org_narrative, org_country_lookup)
    activity_list_unnest_5$country_name2 <- as.character(activity_list_unnest_5$country_name2)
    activity_list_unnest_5$country_name2 <- ifelse(activity_list_unnest_5$country_name2 == "NA", "", activity_list_unnest_5$country_name2)
    
    # check if input is a valid 2-digit country code and then does a lookup to the countrycode_list
    activity_list_unnest_5$country_name <- ifelse(is.na(activity_list_unnest_5$country_code) | nchar(activity_list_unnest_5$country_code) < 2, NA, countrycode_list$name[match(activity_list_unnest_5$country_code,countrycode_list$code)])
    activity_list_unnest_5$country_name <- ifelse(is.na(activity_list_unnest_5$country_name), "", activity_list_unnest_5$country_name)
    
    activity_list_unnest_5[activity_list_unnest_5 == ""] <- NA
    
    # This needs fixing because currently I don't know why it's only picking one of the country names when there might be several 
    activity_list_unnest_5$reporting_org_country <- coalesce(activity_list_unnest_5$country_name, activity_list_unnest_5$country_name2)
    
    activity_list_unnest_5 <- activity_list_unnest_5%>% select(-country_name, -country_name2)

    
    # Add on to org file to save
    orgs_to_save <- activity_list_unnest_5 %>% 
                select(project_id = iati_identifier,
                       organisation_name = reporting_org_narrative,
                       organisation_country = reporting_org_country) %>% 
                mutate(organisation_role = 1) %>% # leading
                unique()
    
    org_names_and_locations_1 <- org_names_and_locations_1 %>%
      rbind(orgs_to_save)
  
# Save to Rdata file
# saveRDS(org_names_and_locations_1, file = "Outputs/org_names_and_locations_1_Jan24update.rds")
org_names_and_locations_1 <- readRDS(file = "Outputs/org_names_and_locations_1_Jan24update.rds")
    
# 6) Unlist and aggregate budget
# activity_list_unnest_6 <- partner_activity_comb %>% 
#   select(-activity_id) %>% 
#   filter(lengths(budget) != 0) %>% 
#   unnest(cols = budget) %>% 
#   select(iati_identifier, 
#          budget_status = status.name, 
#          amount = value.value, 
#          currency = value.currency.code,
#          period_start,
#          period_end) 
#  
#       # Find activities with multiple budgets for same period (i.e. indicative and committed)
#         multiple_budgets <- activity_list_unnest_6 %>% 
#         select(iati_identifier, budget_status, period_start, period_end) %>%
#         unique() %>% 
#         group_by(iati_identifier, period_start, period_end) %>% 
#         summarise(count = n()) %>% 
#         filter (count > 1)
#       
#       # Keep only the committed budget in these cases
#         activity_list_unnest_6 <- activity_list_unnest_6 %>% 
#         filter(!(iati_identifier %in% multiple_budgets$iati_identifier) |
#                  budget_status == "Committed")
#       
#       # Sum to get total budget per activity
#         activity_list_unnest_6 <- activity_list_unnest_6 %>% 
#         group_by(iati_identifier, currency) %>% 
#         summarise(period_start = min(period_start),
#                   period_end = max(period_end),
#                   amount = sum(amount))


activity_list_unnest_6_test <- partner_activity_comb %>% 
  select(iati_identifier, 
         reporting_org_narrative,
         iati_identifier, budget_status,budget_value,
         budget_value_currency,budget_period_start_iso_date,
         budget_period_end_iso_date) %>%
          filter(lengths(budget_value) != 0) %>%
  unique()

activity_list_unnest_6_test$len_val <- lengths(activity_list_unnest_6_test$budget_value)
activity_list_unnest_6_test$len_curr <- lengths(activity_list_unnest_6_test$budget_value_currency)
activity_list_unnest_6_test$diff <- activity_list_unnest_6_test$len_val - activity_list_unnest_6_test$len_curr

activity_list_unnest_6_test_curr <- activity_list_unnest_6_test %>%
  select(iati_identifier, budget_value_currency) %>%
  unnest(c(budget_value_currency))

activity_list_unnest_6_test_curr_unique <- activity_list_unnest_6_test_curr %>% 
  unique() %>%
  group_by(iati_identifier) %>%
  summarise(budget_value_currency = paste(budget_value_currency, collapse = ", "))

# separate out those with multiple currencies
activity_list_unnest_6_test_curr_multi <- activity_list_unnest_6_test_curr[activity_list_unnest_6_test_curr$iati_identifier == "GB-COH-02622374-energy_saving_trust",]
activity_list_unnest_6_test_curr_single <- activity_list_unnest_6_test_curr[activity_list_unnest_6_test_curr$iati_identifier != "GB-COH-02622374-energy_saving_trust",]

activity_list_unnest_6_test_val <- activity_list_unnest_6_test %>%
  select(iati_identifier, reporting_org_narrative, budget_status,budget_value,budget_period_start_iso_date,budget_period_end_iso_date) %>%
  unnest(c(budget_status,budget_value,budget_period_start_iso_date,budget_period_end_iso_date)) %>%
  unique()

activity_list_unnest_6_test_val_multi <- activity_list_unnest_6_test_val[activity_list_unnest_6_test_val$iati_identifier == "GB-COH-02622374-energy_saving_trust",]
activity_list_unnest_6_test_val_single <- activity_list_unnest_6_test_val[activity_list_unnest_6_test_val$iati_identifier != "GB-COH-02622374-energy_saving_trust",]


activity_list_unnest_6_single <- activity_list_unnest_6_test_val_single %>% 
    left_join(activity_list_unnest_6_test_curr_single, by = "iati_identifier") %>%
  unique()

activity_list_unnest_6_test_val_multi$budget_value_currency <- activity_list_unnest_6_test_curr_multi$budget_value_currency

activity_list_unnest_6 <- rbind(activity_list_unnest_6_single, activity_list_unnest_6_test_val_multi)

activity_list_unnest_6$budget_value_currency <- ifelse(is.na(activity_list_unnest_6$budget_value_currency),"",activity_list_unnest_6$budget_value_currency)

# there are some blanks in the currency fields, i am using the below to see what currencies organisations are reporting under then will apply this to the nulls (will have to be a manual job)

# budget_curr <- activity_list_unnest_6 %>% select(reporting_org_narrative, budget_value_currency) %>% unnest(budget_value_currency) %>% unique()


# fixing the currencies manually, will need to be checking this in future iterations in case it changes

GBP_orgs <- c("London School of Hygiene and Tropical Medicine","Social Development Direct Limited","The Institute of Development Studies","Centre for Economic Policy Research",
              "PATH","Palladium","The University of Manchester","Tetra Tech International Development UK Ltd","University of Leeds","Save The Children UK","ODI",
              "IMC WORLDWIDE LTD","SOAS University of London","The University of Oxford")
EUR_orgs <- c("Tetra Tech International Development B.V.")
USD_orgs <- c("AgResults","International Labour Organization (ILO)","CGIAR")

activity_list_unnest_6$budget_value_currency <- ifelse(activity_list_unnest_6$reporting_org_narrative %in% GBP_orgs, "GBP", activity_list_unnest_6$budget_value_currency)
activity_list_unnest_6$budget_value_currency <- ifelse(activity_list_unnest_6$reporting_org_narrative %in% EUR_orgs, "EUR", activity_list_unnest_6$budget_value_currency)
activity_list_unnest_6$budget_value_currency <- ifelse(activity_list_unnest_6$reporting_org_narrative %in% USD_orgs, "USD", activity_list_unnest_6$budget_value_currency)

budget_curr2 <- activity_list_unnest_6 %>% select(reporting_org_narrative, budget_value_currency) %>% unnest(budget_value_currency) %>% unique()

        # Find activities with multiple budgets for same period (i.e. indicative and committed)
        multiple_budgets <- activity_list_unnest_6 %>% 
          select(iati_identifier, budget_status, budget_period_start_iso_date, budget_period_end_iso_date) %>%
          unique() %>% 
          group_by(iati_identifier, budget_period_start_iso_date, budget_period_end_iso_date) %>% 
          summarise(count = n()) %>% 
          filter (count > 1)
        
        # Keep only the committed budget in these cases
        activity_list_unnest_6 <- activity_list_unnest_6 %>% 
          filter(!(iati_identifier %in% multiple_budgets$iati_identifier) |
                   budget_status == "2")
        
        # Sum to get total budget per activity
        activity_list_unnest_6 <- activity_list_unnest_6 %>% 
          select(-reporting_org_narrative) %>%
          group_by(iati_identifier, budget_value_currency) %>% 
          summarise(budget_period_start_iso_date = min(budget_period_start_iso_date),
                    budget_period_end_iso_date = max(budget_period_end_iso_date),
                    budget_value = sum(budget_value)) %>%
          rename(period_start = budget_period_start_iso_date, period_end = budget_period_end_iso_date, amount = budget_value, currency = budget_value_currency)
 

        
# 7) Unlist start/end dates
# activity_list_unnest_7 <- partner_activity_comb %>% 
#   unnest(cols = activity_date) %>% 
#   select(iati_identifier, 
#          date = iso_date,
#          date_type = type.name) %>%
#   # take the first date in cases of two of the same time
#   group_by(iati_identifier, date_type) %>%
#   slice(1) %>%
#   spread(key = date_type, value = date) %>% 
#   mutate(start_date = coalesce(`Actual start`, `Planned start`),
#          end_date = coalesce(`Actual end`, `Planned End`)) %>% 
#   select(iati_identifier, start_date, end_date)

activity_list_unnest_7 <- partner_activity_comb %>% 
  #filter(lengths(activity_date_iso_date) != 0) %>%
  unnest(cols = c(activity_date_iso_date,activity_date_type)) %>% 
  select(iati_identifier, 
         date = activity_date_iso_date,
         date_type = activity_date_type) %>%
  group_by(iati_identifier, date_type) %>% 
  mutate(date_type = str_replace(date_type, "1", "Planned start"),
         date_type = str_replace(date_type, "2", "Actual start"),
         date_type = str_replace(date_type, "3", "Planned end"),
         date_type = str_replace(date_type, "4", "Actual end")) %>%
  slice(1) %>%
  spread(key = date_type, value = date) %>% 
  mutate(start_date = coalesce(`Actual start`, `Planned start`),
         end_date = coalesce(`Actual end`, `Planned end`)) %>% 
  select(iati_identifier, start_date, end_date)



# 8) Extract transactions

##### this extracts countries listed in the transactions####

specific_trans_extract_country <- function(id) {
  page_list <- list()
  page <- 1
  page_df = data.frame()
  
  while (!is.null(page_df)) {
    Sys.sleep(1)
    message(page)
    page_df <- transactions_extract_country(page, id)
    if(!is.null(page_df)){
      page_list[[page]] = page_df
    }
    page = page + 1
  }
  rbindlist(page_list, fill=T)
}

uk_activity_ids = unique(partner_activity_comb$iati_identifier)
uk_activity_ids <- URLencode(uk_activity_ids)

# My batches code isn't working at the moment, so having to run them individually
# batch_size = 15
# batches = c()
# current_batch = c()
# for(i in 1:length(uk_activity_ids)){
#   current_id = uk_activity_ids[i]
#   if(i %% batch_size == 0){
#     current_batch_str = paste0('("', paste(current_batch, collapse = '" OR "'), '")')
#     batches = c(batches, current_batch_str)
#     current_batch = c(current_id)
#   } else {
#     current_batch = c(current_batch, current_id)
#   }
# }

# transaction_list_countries <- lapply(batches, specific_trans_extract_country)
transaction_list_countries <- lapply(uk_activity_ids, specific_trans_extract_country)
transaction_list_countries = rbindlist(transaction_list_countries, fill=TRUE)


# Save to Rdata file
# saveRDS(transaction_list_countries, file = "Outputs/transaction_list_countries_Jan24update.rds")
transaction_list_countries <- readRDS(file = "Outputs/transaction_list_countries_Jan24update.rds")

transaction_list_countries$recipient_country <- countrycode_list$name[match(transaction_list_countries$transaction_recipient_country_code,countrycode_list$code)]

    # Extract recipient countries (where included in transactions)
        transaction_countries <- transaction_list_countries %>% 
          select(iati_identifier, recipient_country) %>% 
            # rename and remove blanks
          filter(!is.na(recipient_country)) %>% 
          unique()
        
        # Summarise countries for joining to main dataset
        transaction_countries_summarised <- transaction_countries %>% 
          group_by(iati_identifier) %>% 
          summarise(recipient_country = paste(coalesce(recipient_country, ""), collapse = ", "))
        

##### this extracts recipients
        
specific_trans_extract_recipient <- function(id) {
          page_list <- list()
          page <- 1
          page_df = data.frame()
          
          while (!is.null(page_df)) {
            Sys.sleep(1)
            message(page)
            page_df <- transactions_extract_recipient(page, id)
            if(!is.null(page_df)){
              page_list[[page]] = page_df
            }
            page = page + 1
          }
          rbindlist(page_list, fill=T)
        }
        
uk_activity_ids = unique(partner_activity_comb$iati_identifier)
uk_activity_ids <- URLencode(uk_activity_ids)

# My batches code isn't working at the moment, so having to run them individually
# batch_size = 15
# batches = c()
# current_batch = c()
# for(i in 1:length(uk_activity_ids)){
#   current_id = uk_activity_ids[i]
#   if(i %% batch_size == 0){
#     current_batch_str = paste0('("', paste(current_batch, collapse = '" OR "'), '")')
#     batches = c(batches, current_batch_str)
#     current_batch = c(current_id)
#   } else {
#     current_batch = c(current_batch, current_id)
#   }
# }

# transaction_list_recipient <- lapply(batches, specific_trans_extract_recipient)
transaction_list_recipient <- lapply(uk_activity_ids, specific_trans_extract_recipient)
transaction_list_recipient = rbindlist(transaction_list_recipient, fill=TRUE)


# Save to Rdata file
# saveRDS(transaction_list_recipient, file = "Outputs/transaction_list_recipient_Jan24update.rds")
transaction_list_recipient <- readRDS(file = "Outputs/transaction_list_recipient_Jan24update.rds")

        
    # Extract receiver organisations
        transaction_receiver_orgs <- transaction_list_recipient %>% 
          select(iati_identifier, transaction_receiver_org_narrative) %>% 
            # unnest organisation names
          #filter(lengths(receiver_organisation.narrative) != 0) %>% 
          unnest(cols = transaction_receiver_org_narrative) %>% 
          #select(-lang.code, -lang.name) %>% 
          #rename(transaction_receiver_name = text) %>% 
            # Exclude blanks, other text
          filter(!is.na(transaction_receiver_org_narrative), 
                 !str_detect(str_to_lower(transaction_receiver_org_narrative), "reimbursement"),
                 !str_detect(str_to_lower(transaction_receiver_org_narrative), "disbursement")) %>% 
          unique()
        
            # Add to organisation name and country database
            receiver_orgs_to_save <- transaction_receiver_orgs %>% 
              inner_join(no_partner_info, by = "iati_identifier") %>% 
              rename(project_id = iati_identifier,
                     organisation_name = transaction_receiver_org_narrative) %>% 
              # Look up country from both country code and organisation name
              mutate(organisation_country = map(organisation_name, org_country_lookup)) %>% 
              mutate(organisation_country = unlist(organisation_country)) %>%
              mutate(organisation_role = 2) # partners
              
              # Add on to org file to save
              org_names_and_locations_1 <- org_names_and_locations_1 %>% 
              rbind(receiver_orgs_to_save)      
       
              unique(transaction_receiver_orgs$transaction_receiver_org_narrative)        
        
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "**********", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "*****", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)             
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Direct expenditure, multiple suppliers", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Partner country based NGO", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)              
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "International NGO", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)              
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Govermental", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)              
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Multilateral", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)              
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Aug'18 Payment", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)              
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Multiple Project Consultants", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)              
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Total other Project expenses", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)   
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Multiple partners paid through expenditure", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)   
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "none", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)   
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Options", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)   
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Various Suppliers", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative)   
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "General partner transfer", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative) 
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Divers fournisseurs", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative) 
transaction_receiver_orgs$transaction_receiver_org_narrative <- 
  ifelse(transaction_receiver_orgs$transaction_receiver_org_narrative == "Not disclosed.", "", 
         transaction_receiver_orgs$transaction_receiver_org_narrative) 

              
        # Summarise orgs for joining to main dataset
        transaction_orgs_summarised <- transaction_receiver_orgs %>% 
          group_by(iati_identifier) %>% 
          summarise(transaction_receiver_org_narrative = paste(coalesce(transaction_receiver_org_narrative, ""), collapse = ", "))
       
         # Summarise org countries for joining to main dataset
        transaction_org_countries_summarised <- receiver_orgs_to_save %>% 
          select(iati_identifier = project_id, organisation_country) %>% 
          unique() %>% 
          filter(!is.na(organisation_country)) %>% 
          group_by(iati_identifier) %>% 
          summarise(transaction_receiver_country = paste(coalesce(organisation_country, ""), collapse = ", "))
        
      
      # Join on transactions country and org info to relevant datasets
        activity_list_unnest_2 <- partner_activity_comb %>% 
          select(iati_identifier) %>% 
          left_join(activity_list_unnest_2, by = "iati_identifier") %>%
          left_join(transaction_countries_summarised, by = "iati_identifier") %>% 
          mutate(recipient_country = coalesce(recipient_country_region, recipient_country)) %>% 
          select(-recipient_country_region)

        
        activity_list_unnest_4 <- partner_activity_comb %>% 
          select(iati_identifier) %>% 
          left_join(activity_list_unnest_4, by = "iati_identifier") %>%
          left_join(transaction_orgs_summarised, by = "iati_identifier") %>% 
          left_join(transaction_org_countries_summarised, by = "iati_identifier") %>% 
          mutate(partner = coalesce(partner, transaction_receiver_org_narrative),
                 partner_country = coalesce(partner_country, transaction_receiver_country)) %>% 
          select(-transaction_receiver_org_narrative, -transaction_receiver_country) %>%
          unique()

        # haven't figured out what activity_list_unnest_8 does, due to lack of my comments
        
# activity_list_unnest_8 <- partner_activity_comb %>%
#           select(iati_identifier, participating_org_ref, participating_org_narrative, participating_org_activity_id) %>%
#           unnest(cols = c(participating_org_ref, participating_org_narrative, participating_org_activity_id)) %>%
#           filter(str_detect(participating_org_activity_id, "GB-1-|GB-GOV-1")) %>%
#           select(iati_identifier, participating_org_activity_id) %>%
#           unique()

# Join unnested info to original data
activity_list <- activity_list_base %>% 
  left_join(activity_list_unnest_1, by = "iati_identifier") %>%
  left_join(activity_list_unnest_2, by = "iati_identifier") %>%
  left_join(activity_list_unnest_3, by = "iati_identifier") %>%
  left_join(activity_list_unnest_4, by = "iati_identifier") %>% 
  left_join(activity_list_unnest_5, by = "iati_identifier") %>% 
  left_join(activity_list_unnest_6, by = "iati_identifier") %>% 
  left_join(activity_list_unnest_7, by = "iati_identifier")


# Assign a reporting org name if missing 
activity_list <- activity_list %>% 
  mutate(reporting_org = coalesce(reporting_org_narrative, reporting_org_ref, gov_funder))

# Reorder columns and add date of refresh
activity_list <- activity_list %>% 
  select(iati_identifier, reporting_org_ref, reporting_org, reporting_org_country,
         hierarchy, activity_status, activity_id,
         activity_title, activity_description, start_date, end_date,
         recipient_country, sector_name,
         partner, partner_country, 
         gov_funder, fund,
         amount, period_start, period_end, currency) %>% 
  unique() %>% 
  mutate(refresh_date = Sys.Date())



activity_list <- activity_list %>%  
      # Add missing FCDO activity IDs (IDRC and AgResults)
    rename(programme_id = activity_id) %>% 
    mutate(programme_id = case_when(str_detect(iati_identifier, "XM-DAC-301-2") & str_detect(activity_description, "CLARE") ~ "GB-GOV-1-300126",
                                    str_detect(iati_identifier, "XM-DAC-301-2") & str_detect(activity_description, "CARIAA") ~ "GB-1-203506", 
                                    str_detect(iati_identifier, "XI-IATI-AGR") ~ "GB-1-203052",
                                    TRUE ~ programme_id),  
      # Add Fund and Funder label  
           fund = case_when(
                    str_detect(programme_id, "GB-GOV-1-|GB-1-") ~ "FCDO Research - Programmes",
                    str_detect(programme_id, "GB-GOV-10") ~ "DHSC - Global Health Research - Partnerships",
                    TRUE ~ fund),
           gov_funder = case_when(
             str_detect(programme_id, "GB-GOV-1-|GB-1-") ~ "Foreign, Commonwealth and Development Office",
             str_detect(programme_id, "GB-GOV-10") ~ "Department of Health and Social Care",
             TRUE ~ gov_funder)) %>% unique()

# Save to Rdata file
# saveRDS(activity_list, file = "Outputs/partner_activity_list_Jan24update_v2.rds")
activity_list <- readRDS(file = "Outputs/partner_activity_list_Jan24update_v2.rds")
# write.xlsx(activity_list, file = "Outputs/partner_activity_list_kp.xlsx")

# Save org names and countries to file
saveRDS(org_names_and_locations_1, file = "Outputs/org_names_and_locations_1_Jan24update.rds")

# Clear environment
rm(partner_activity_extract, partnership_activities, partner_activities_via_title, partner_activities_via_funder,
   result, new_rows, x, y, page, ri_linked_activites,
   activity_list_base, activity_list_unnest_1, activity_list_unnest_2, activity_list_unnest_3, activity_list_unnest_4,
   activity_list_unnest_4_partner_countries, activity_list_unnest_4_partner_names, activity_list_unnest_5,
   activity_list_unnest_6, activity_list_unnest_7, unlinked_partner_iati_activity_ids)
